+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### plt
 
+- Layouts for setting subplot areas
 - `visible` member of SubplotDescriptor to control the visibility of an axis line.
 
 ### Changed
@@ -19,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Split `Ticker` into separate `TickSpacing` and `TickLabels` structs.
 - `Auto` varients of `TickSpacing` and `TickLabels` now determine if they should exist
   based on whether a plot uses that axis.
+- Deprecated `Figure::add_subplot` in favor of using layouts.
 
 ## [0.2.1] - 2022-09-02